refactor: Add custom hook to fetch product inventory
