+++ src/templates/ProductPage.js
@@ -1,68 +1,16 @@
-import React, { useContext, useEffect, useReducer, useState } from 'react'
+import React from 'react'
 import { graphql, withPrefix } from 'gatsby'
 
 import SEO from '../components/SEO'
 import Photo from '../components/Photo'
 import Badge from '../components/Badge'
 import AddToCart from '../components/AddToCart'
-import { Shopkit } from '../shopkit'
-
-function reducer(state, { type, inventory }) {
-  switch (type) {
-    case 'IN_STOCK':
-      return {
-        inStock: true,
-        ...inventory
-      }
-    case 'OUT_OF_STOCK':
-      return {
-        inStock: false,
-        ...inventory
-      }
-    default:
-      return state
-  }
-}
+import useMoltinInventory from '../hooks/useMoltinInventory'
 
 function ProductPage({ data: { product } }) {
-  const { moltin } = useContext(Shopkit)
-  const [state, dispatch] = useReducer(reducer, {
-    inStock: null,
-    available: 0
-  })
-  const [inventoryLoading, setInventoryLoading] = useState(true)
-  const [inventoryError, setInventoryError] = useState(false)
-
-  async function getProductInventory() {
-    try {
-      const {
-        data: { available }
-      } = await moltin.get(`inventories/${product.id}`)
-
-      dispatch({
-        type: available === 0 ? 'OUT_OF_STOCK' : 'IN_STOCK',
-        inventory: { available }
-      })
-      setInventoryLoading(false)
-    } catch ({ errors: [error] }) {
-      console.error(error)
-      setInventoryError(
-        `There was a problem retrieving the inventory details for ${
-          product.name
-        }`
-      )
-    }
-  }
-
-  useEffect(() => {
-    if (!product.manage_stock) {
-      dispatch({ type: 'IN_STOCK' })
-
-      return setInventoryLoading(false)
-    }
-
-    getProductInventory()
-  }, [])
+  const [inventory, inventoryLoading, inventoryError] = useMoltinInventory(
+    product
+  )
 
   const {
     meta: { display_price }
@@ -105,12 +53,12 @@ function ProductPage({ data: { product } }) {
                   <Badge className="mx-2"> Loading inventory</Badge>
                 ) : (
                   <Badge
-                    color={state.inStock ? 'green' : 'red'}
+                    color={inventory.inStock ? 'green' : 'red'}
                     className="mx-2"
                   >
-                    {state.inStock
+                    {inventory.inStock
                       ? product.manage_stock
-                        ? `${state.available} in stock `
+                        ? `${inventory.available} in stock `
                         : 'In Stock'
                       : 'Out of stock'}
                   </Badge>
@@ -122,7 +70,7 @@ function ProductPage({ data: { product } }) {
             inventoryError
           ) : (
             <div className="flex flex-wrap flex-col md:flex-row md:items-end">
-              <AddToCart productId={product.id} disabled={!state.inStock} />
+              <AddToCart productId={product.id} disabled={!inventory.inStock} />
             </div>
           )}
